# Perf improvements by lazy loading some dependencies