## Summary

I know that it is had to benchmark small improvements, also considering all the other things that are happening in a computer while a run the tests and that can interfere with the timings.

With a warm cache I consistently see an improvement of around 2-3 seconds when running  jest packages in the facebook/jest repo. Before: ~23 seconds, After ~20/21 seconds

I see similar results with other projects. For example in webpack/webpack I consistently see around a half a second improvement when running jest unittest Before: ~3.0 seconds, After: ~2.5 seconds

I know that this small improvements are really hard to deterministically measure, which is why I wanted to see if it even shows an improvement in other computers, other projects or different OS(Windows)

## Test plan

I created this small tool to benchmark it `npx jest-benchmarks`(https://github.com/rogeliog/jest-benchmark) to benchmark it.

```
`jest jest-circus`
# Before
Number of runs: 10
Average: 2.896s
Max: 3.092s
Min: 2.574s

# After:
Number of runs: 10
Average: 2.392s
Max: 2.723s
Min: 2.114s
```

```
`jest jest-r`

# Before
Number of runs: 10
Average: 8.025s
Max: 9.128s
Min: 7.209s

# After
Number of runs: 10
Average: 6.701
Max: 7.505
Min: 6.251
```