@rogeliog 2018-10-18
@SimenB I agree, that is a good point. @thymikee brought up a similar point yesterday 😄 	


As you mentioned most of the dependencies that I inlined were inlined by the babel transformation. What I'm the most confused about is that by manually inlining them I still see a performance improvement compared to master. This is why I posted the other day asking for people to test in their own computers, because I've been trying to find and understand the root cause without any luck

As @thymikee suggested, I compared the build output of each of the modified files against the build output in master_(see the bottom of this comment for a diff)_. For all of these dependencies, execept `babel-types` the build process was already inlining them.

Here are some runs of `npx jest-benchmarks`. As you can see, specifically for the second benchmark, I've been consistently getting better numbers by manualy inlining all of those deps.


- `#Master`: benchmark in the master branch
- `#With just babel-types`: benchmark in a branch were I only inlined babel-types
- `#Everything`: benchmark in this branch.


```
# Master

Command: jest jest-circus
Number of runs: 10
Average: 2.453s
Max: 2.574s
Min: 2.349s

# With just babel-types

Command: jest jest-circus
Number of runs: 10
Average: 2.218s
Max: 2.477s
Min: 2.07s

# Everything

Command: jest jest-circus
Number of runs: 10
Average: 2.132s
Max: 2.082s
Min: 2.349s

```

```
# Master
Command: jest jest-r
Number of runs: 10
Average: 8.379s
Max: 10.666s
Min: 8.011s

# With just babel-types
Command: jest jest-r
Number of runs: 10
Average: 8.352s
Max: 9.32s
Min: 7.859s

# Everything
Command: jest jest-r
Number of runs: 10
Average: 6.829s
Max: 7.133s
Min: 6.546s
```

## build files

_Note: Ignore the `build-files` dir, I created it in a branch just for comparing this files_

```diff
diff --git a/build-files/babel-jest-index-master.js b/build-files/babel-jest-index.js
index bf7325378..84fabb268 100644
--- a/build-files/babel-jest-index-master.js
+++ b/build-files/babel-jest-index.js
@@ -32,18 +32,11 @@ function _load_babelCore() {
   return (_babelCore = require('babel-core'));
 }

-var _babelPluginIstanbul;
-
-function _load_babelPluginIstanbul() {
-  return (_babelPluginIstanbul = _interopRequireDefault(
-    require('babel-plugin-istanbul')
-  ));
-}
-
 function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }

+const getBabelPluginIstanbul = () => require('babel-plugin-istanbul').default;
 /**
  * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
  *
@@ -174,11 +167,12 @@ const createTransformer = options => {

       const theseOptions = Object.assign({filename: filename}, options);
       if (transformOptions && transformOptions.instrument) {
+        const babelIstanbulPlugin = getBabelPluginIstanbul();
         theseOptions.auxiliaryCommentBefore = ' istanbul ignore next ';
         // Copied from jest-runtime transform.js
         theseOptions.plugins = theseOptions.plugins.concat([
           [
-            (_babelPluginIstanbul || _load_babelPluginIstanbul()).default,
+            babelIstanbulPlugin,
             {
               // files outside `cwd` will not be instrumented
               cwd: config.rootDir,



diff --git a/build-files/jest-edit-support-Snapshot-master.js b/build-files/jest-edit-support-Snapshot.js
index af4d8cdaa..1430cf243 100644
--- a/build-files/jest-edit-support-Snapshot-master.js
+++ b/build-files/jest-edit-support-Snapshot.js
@@ -14,12 +14,6 @@ Object.defineProperty(exports, '__esModule', {
   value: true
 });

-var _babelTraverse;
-
-function _load_babelTraverse() {
-  return (_babelTraverse = _interopRequireDefault(require('babel-traverse')));
-}
-
 var _babylon_parser;

 function _load_babylon_parser() {
@@ -32,10 +26,6 @@ function _load_jestSnapshot() {
   return (_jestSnapshot = require('jest-snapshot'));
 }

-function _interopRequireDefault(obj) {
-  return obj && obj.__esModule ? obj : {default: obj};
-}
-
 const describeVariants = Object.assign(Object.create(null), {
   describe: true,
   fdescribe: true,
@@ -106,6 +96,7 @@ class Snapshot {
   }

   getMetadata(filePath) {
+    const traverse = require('babel-traverse').default;
     const fileNode = this._parser(filePath);
     const state = {
       found: []
@@ -121,7 +112,7 @@ class Snapshot {
       }
     };

-    (0, (_babelTraverse || _load_babelTraverse()).default)(fileNode, {
+    traverse(fileNode, {
       enter: path => {
         const visitor = Visitors[path.node.type];
         if (visitor != null) {



diff --git a/build-files/jest-haste-map-index-master.js b/build-files/jest-haste-map-index.js
index 92da6055f..4b4e60689 100644
--- a/build-files/jest-haste-map-index-master.js
+++ b/build-files/jest-haste-map-index.js
@@ -152,12 +152,6 @@ function _load_path() {
   return (_path = _interopRequireDefault(require('path')));
 }

-var _sane;
-
-function _load_sane() {
-  return (_sane = _interopRequireDefault(require('sane')));
-}
-
 var _jestSerializer;

 function _load_jestSerializer() {
@@ -210,6 +204,8 @@ function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }

+const getSane = () => require('sane').default;
+
 const CHANGE_INTERVAL = 30;
 const MAX_WAIT_TIME = 240000;
 const NODE_MODULES =
@@ -873,13 +869,16 @@ class HasteMap extends (_events || _load_events()).default {
     // all files, even changes to node_modules.
     this._options.throwOnModuleCollision = false;
     this._options.retainAllFiles = true;
-
-    const Watcher =
-      canUseWatchman && this._options.useWatchman
-        ? (_watchman_watcher || _load_watchman_watcher()).default
-        : (_os || _load_os()).default.platform() === 'darwin'
-          ? (_sane || _load_sane()).default.FSEventsWatcher
-          : (_sane || _load_sane()).default.NodeWatcher;
+    let Watcher;
+    if (canUseWatchman && this._options.useWatchman) {
+      Watcher = (_watchman_watcher || _load_watchman_watcher()).default;
+    } else {
+      const sane = getSane();
+      Watcher =
+        (_os || _load_os()).default.platform() === 'darwin'
+          ? sane.FSEventsWatcher
+          : sane.NodeWatcher;
+    }
     const extensions = this._options.extensions;
     const ignorePattern = this._options.ignorePattern;
     const rootDir = this._options.rootDir;



diff --git a/build-files/jest-runtime-script-transformer-master.js b/build-files/jest-runtime-script-transformer.js
index 49d27e178..f9ade2019 100644
--- a/build-files/jest-runtime-script-transformer-master.js
+++ b/build-files/jest-runtime-script-transformer.js
@@ -34,20 +34,6 @@ function _load_gracefulFs() {
   return (_gracefulFs = _interopRequireDefault(require('graceful-fs')));
 }

-var _babelCore;
-
-function _load_babelCore() {
-  return (_babelCore = require('babel-core'));
-}
-
-var _babelPluginIstanbul;
-
-function _load_babelPluginIstanbul() {
-  return (_babelPluginIstanbul = _interopRequireDefault(
-    require('babel-plugin-istanbul')
-  ));
-}
-
 var _convertSourceMap;

 function _load_convertSourceMap() {
@@ -114,12 +100,7 @@ function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }

-// This data structure is used to avoid recalculating some data every time that
-// we need to transform a file. Since ScriptTransformer is instantiated for each
-// file we need to keep this object in the local scope of this module.
-const projectCaches = new WeakMap();
-
-// To reset the cache for specific changesets (rather than package version).
+const getBabelCore = () => require('babel-core');
 /**
  * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
  *
@@ -129,6 +110,14 @@ const projectCaches = new WeakMap();
  *
  */

+const getBabelPluginIstanbul = () => require('babel-plugin-istanbul').default;
+
+// This data structure is used to avoid recalculating some data every time that
+// we need to transform a file. Since ScriptTransformer is instantiated for each
+// file we need to keep this object in the local scope of this module.
+const projectCaches = new WeakMap();
+
+// To reset the cache for specific changesets (rather than package version).
 const CACHE_VERSION = '1';

 class ScriptTransformer {
@@ -250,13 +239,18 @@ class ScriptTransformer {
   }

   _instrumentFile(filename, content) {
-    return (0, (_babelCore || _load_babelCore()).transform)(content, {
+    var _getBabelCore = getBabelCore();
+
+    const babelTransform = _getBabelCore.transform;
+
+    const babelPluginIstanbul = getBabelPluginIstanbul();
+    return babelTransform(content, {
       auxiliaryCommentBefore: ' istanbul ignore next ',
       babelrc: false,
       filename: filename,
       plugins: [
         [
-          (_babelPluginIstanbul || _load_babelPluginIstanbul()).default,
+          babelPluginIstanbul,
           {
             compact: false,
             // files outside `cwd` will not be instrumented
fatal: Path 'build-files/jest-snapshot-inline-snapshot.jsgit' does not exist in 'head'



diff --git a/build-files/jest-edit-support-Snapshot-master.js b/build-files/jest-edit-support-Snapshot.js
index af4d8cdaa..1430cf243 100644
--- a/build-files/jest-edit-support-Snapshot-master.js
+++ b/build-files/jest-edit-support-Snapshot.js
@@ -14,12 +14,6 @@ Object.defineProperty(exports, '__esModule', {
   value: true
 });

-var _babelTraverse;
-
-function _load_babelTraverse() {
-  return (_babelTraverse = _interopRequireDefault(require('babel-traverse')));
-}
-
 var _babylon_parser;

 function _load_babylon_parser() {
@@ -32,10 +26,6 @@ function _load_jestSnapshot() {
   return (_jestSnapshot = require('jest-snapshot'));
 }

-function _interopRequireDefault(obj) {
-  return obj && obj.__esModule ? obj : {default: obj};
-}
-
 const describeVariants = Object.assign(Object.create(null), {
   describe: true,
   fdescribe: true,
@@ -106,6 +96,7 @@ class Snapshot {
   }

   getMetadata(filePath) {
+    const traverse = require('babel-traverse').default;
     const fileNode = this._parser(filePath);
     const state = {
       found: []
@@ -121,7 +112,7 @@ class Snapshot {
       }
     };

-    (0, (_babelTraverse || _load_babelTraverse()).default)(fileNode, {
+    traverse(fileNode, {
       enter: path => {
         const visitor = Visitors[path.node.type];
         if (visitor != null) {



diff --git a/build-files/jest-haste-map-index-master.js b/build-files/jest-haste-map-index.js
index 92da6055f..4b4e60689 100644
--- a/build-files/jest-haste-map-index-master.js
+++ b/build-files/jest-haste-map-index.js
@@ -152,12 +152,6 @@ function _load_path() {
   return (_path = _interopRequireDefault(require('path')));
 }

-var _sane;
-
-function _load_sane() {
-  return (_sane = _interopRequireDefault(require('sane')));
-}
-
 var _jestSerializer;

 function _load_jestSerializer() {
@@ -210,6 +204,8 @@ function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }

+const getSane = () => require('sane').default;
+
 const CHANGE_INTERVAL = 30;
 const MAX_WAIT_TIME = 240000;
 const NODE_MODULES =
@@ -873,13 +869,16 @@ class HasteMap extends (_events || _load_events()).default {
     // all files, even changes to node_modules.
     this._options.throwOnModuleCollision = false;
     this._options.retainAllFiles = true;
-
-    const Watcher =
-      canUseWatchman && this._options.useWatchman
-        ? (_watchman_watcher || _load_watchman_watcher()).default
-        : (_os || _load_os()).default.platform() === 'darwin'
-          ? (_sane || _load_sane()).default.FSEventsWatcher
-          : (_sane || _load_sane()).default.NodeWatcher;
+    let Watcher;
+    if (canUseWatchman && this._options.useWatchman) {
+      Watcher = (_watchman_watcher || _load_watchman_watcher()).default;
+    } else {
+      const sane = getSane();
+      Watcher =
+        (_os || _load_os()).default.platform() === 'darwin'
+          ? sane.FSEventsWatcher
+          : sane.NodeWatcher;
+    }
     const extensions = this._options.extensions;
     const ignorePattern = this._options.ignorePattern;
     const rootDir = this._options.rootDir;



diff --git a/build-files/jest-runtime-script-transformer-master.js b/build-files/jest-runtime-script-transformer.js
index 49d27e178..f9ade2019 100644
--- a/build-files/jest-runtime-script-transformer-master.js
+++ b/build-files/jest-runtime-script-transformer.js
@@ -34,20 +34,6 @@ function _load_gracefulFs() {
   return (_gracefulFs = _interopRequireDefault(require('graceful-fs')));
 }

-var _babelCore;
-
-function _load_babelCore() {
-  return (_babelCore = require('babel-core'));
-}
-
-var _babelPluginIstanbul;
-
-function _load_babelPluginIstanbul() {
-  return (_babelPluginIstanbul = _interopRequireDefault(
-    require('babel-plugin-istanbul')
-  ));
-}
-
 var _convertSourceMap;

 function _load_convertSourceMap() {
@@ -114,12 +100,7 @@ function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }

-// This data structure is used to avoid recalculating some data every time that
-// we need to transform a file. Since ScriptTransformer is instantiated for each
-// file we need to keep this object in the local scope of this module.
-const projectCaches = new WeakMap();
-
-// To reset the cache for specific changesets (rather than package version).
+const getBabelCore = () => require('babel-core');
 /**
  * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
  *
@@ -129,6 +110,14 @@ const projectCaches = new WeakMap();
  *
  */

+const getBabelPluginIstanbul = () => require('babel-plugin-istanbul').default;
+
+// This data structure is used to avoid recalculating some data every time that
+// we need to transform a file. Since ScriptTransformer is instantiated for each
+// file we need to keep this object in the local scope of this module.
+const projectCaches = new WeakMap();
+
+// To reset the cache for specific changesets (rather than package version).
 const CACHE_VERSION = '1';

 class ScriptTransformer {
@@ -250,13 +239,18 @@ class ScriptTransformer {
   }

   _instrumentFile(filename, content) {
-    return (0, (_babelCore || _load_babelCore()).transform)(content, {
+    var _getBabelCore = getBabelCore();
+
+    const babelTransform = _getBabelCore.transform;
+
+    const babelPluginIstanbul = getBabelPluginIstanbul();
+    return babelTransform(content, {
       auxiliaryCommentBefore: ' istanbul ignore next ',
       babelrc: false,
       filename: filename,
       plugins: [
         [
-          (_babelPluginIstanbul || _load_babelPluginIstanbul()).default,
+          babelPluginIstanbul,
           {
             compact: false,
             // files outside `cwd` will not be instrumented



diff --git a/build-files/jest-snapshot-inline-snapshot-master.js b/build-files/jest-snapshot-inline-snapshot.js
index e28783073..ef4045b26 100644
--- a/build-files/jest-snapshot-inline-snapshot-master.js
+++ b/build-files/jest-snapshot-inline-snapshot.js
@@ -17,14 +17,23 @@ var _path = require('path');

 var _path2 = _interopRequireDefault(_path);

-var _babelTypes = require('babel-types');
-
 var _utils = require('./utils');

 function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }

+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ *
+ */
+
+const getBabelTypes = () => require('babel-types');
+
 const saveInlineSnapshots = (exports.saveInlineSnapshots = (
   snapshots,
   prettier,
@@ -56,14 +65,6 @@ const saveInlineSnapshots = (exports.saveInlineSnapshots = (
     );
   }
 });
-/**
- * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- *
- */

 const saveSnapshotsForFile = (
   snapshots,
@@ -125,7 +126,13 @@ const createParser = (snapshots, inferredParser, babelTraverse) => (
   parsers,
   options
 ) => {
+  var _getBabelTypes = getBabelTypes();
+
+  const templateElement = _getBabelTypes.templateElement,
+    templateLiteral = _getBabelTypes.templateLiteral,
+    file = _getBabelTypes.file;
   // Workaround for https://github.com/prettier/prettier/issues/3150
+
   options.parser = inferredParser;

   const groupedSnapshots = groupSnapshotsByFrame(snapshots);
@@ -139,7 +146,7 @@ const createParser = (snapshots, inferredParser, babelTraverse) => (

   // Flow uses a 'Program' parent node, babel expects a 'File'.
   if (ast.type !== 'File') {
-    ast = (0, _babelTypes.file)(ast, ast.comments, ast.tokens);
+    ast = file(ast, ast.comments, ast.tokens);
     delete ast.program.comments;
   }

@@ -177,12 +184,8 @@ const createParser = (snapshots, inferredParser, babelTraverse) => (

         remainingSnapshots.delete(snapshot);

-        return (0, _babelTypes.templateLiteral)(
-          [
-            (0, _babelTypes.templateElement)({
-              raw: (0, _utils.escapeBacktickString)(snapshot)
-            })
-          ],
+        return templateLiteral(
+          [templateElement({raw: (0, _utils.escapeBacktickString)(snapshot)})],
           []
         );
       });
```